+++ src/main/java/com/costardstudio/patterns/features/observer/controller/OperationPerform.java
@@ -3,5 +3,5 @@ package com.costardstudio.patterns.features.observer.controller;
 import java.util.List;
 
 public interface OperationPerform {
-    public int computeOperation(List<String> inputs);
+    public float computeOperation(List<String> inputs);
 }
